+++ src/blocks/Chunk.hpp
@@ -10,6 +10,8 @@
 #include "../blocks/Block.hpp"
 #include "../blocks/ChunkMesh.hpp"
 #include "../graphics/PerlinNoise.hpp"
+#include "../blocks/World.hpp"
+#include "../blocks/Vec2Hashes.hpp"
 
 #define CHUNK_X 8
 #define CHUNK_Y 384
@@ -54,5 +56,5 @@ public:
 	BlockType GetBlock(int x, int y, int z) const;
 
 	void GenerateBlockData();
-	void GenerateChunkMesh(const auto& activeChunks);
+	void GenerateChunkMesh(const std::unordered_map<glm::vec2, std::unique_ptr<Chunk>, Vec2Hasher, Vec2Equals>& activeChunks);
 };
